# modules/web-ui/openapi.yaml
openapi: 3.0.3
info:
  title: M6 Web UI Service API (BFF)
  description: |
    **POKER-BRAIN Web UI Service**

    사용자 인터페이스 및 Backend for Frontend (BFF) API입니다.

    **담당**: Frontend Engineer (Frank)
    **모듈 ID**: M6
    **배포**: Cloud Run (us-central1) 또는 Vercel
    **Tech Stack**: Next.js 14 (App Router) + shadcn/ui

    ## 주요 기능
    - 포커 핸드 검색 UI
    - 프록시 영상 미리보기
    - 클립 다운로드 요청 및 상태 조회
    - 즐겨찾기 관리
    - 관리자 대시보드 (Phase 0 진행률)

    ## BFF 패턴
    ```
    Browser (React) → Next.js API Routes (BFF)
        ↓
    BFF → M4 (RAG Search)
    BFF → M5 (Clipping)
    BFF → M3 (Timecode Validation, Admin)
        ↓
    Browser ← JSON Response
    ```

    ## 페이지 구조
    ```
    /                    → 검색 페이지
    /search?q=...        → 검색 결과
    /hand/[hand_id]      → 핸드 상세 + 미리보기
    /downloads           → 내 다운로드 목록
    /favorites           → 즐겨찾기
    /admin               → 관리자 대시보드
    ```

    ## 의존성
    - **Upstream**: M4 (Search), M5 (Clipping), M3 (Admin)
    - **Auth**: Google IAP (자동)

  version: 1.0.0
  contact:
    name: GG Production Frontend Team
    email: aiden.kim@ggproduction.net
servers:
  - url: https://poker-brain-{env}.run.app
    description: Cloud Run deployment
    variables:
      env:
        default: prod
        enum:
          - prod
          - staging
          - dev
  - url: http://localhost:3000
    description: Local development (Next.js)
...
paths:
...
  /api/download:
    post:
      summary: 클립 다운로드 요청 (BFF)
      description: |
        M5 Clipping Service를 호출하여 클립 생성을 요청합니다.

        **BFF 역할**:
        - BigQuery에서 hand 정보 조회
        - M5 API 호출 (POST /v1/clip/request)
        - 다운로드 이력 저장
      operationId: download
      tags:
        - Download
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
            example:
              hand_id: "wsop2024_me_d3_h154"
      responses:
        '200':
          description: 다운로드 요청 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResponse'
              example:
                clip_request_id: "clip-20241117-001"
                status: "queued"
                message: "클립 생성 중입니다. 5분 내 완료됩니다."
                estimated_completion: "2024-11-17T14:05:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/favorites:
    get:
      summary: 즐겨찾기 목록 조회
      operationId: getFavorites
      tags:
        - User
      security:
        - CookieAuth: []
      responses:
        '200':
          description: 즐겨찾기 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/HandSummary'
    post:
      summary: 즐겨찾기 추가
      operationId: addFavorite
      tags:
        - User
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hand_id:
                  type: string
      responses:
        '200':
          description: 추가 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "added"
    delete:
      summary: 즐겨찾기 제거
      operationId: removeFavorite
      tags:
        - User
      security:
        - CookieAuth: []
      parameters:
        - name: hand_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 제거 완료
...
components:
  schemas:
...
    HandSummary:
      type: object
      properties:
        hand_id:
          type: string
        summary:
          type: string
        event_name:
          type: string
        timestamp_start:
          type: string
          format: date-time
        proxy_url:
          type: string
          nullable: true
        players:
          type: array
          items:
            type: string
        pot_size_usd:
          type: number
        is_favorite:
          type: boolean
          description: 현재 사용자의 즐겨찾기 여부
    DownloadRequest:
      type: object
      required:
        - hand_id
      properties:
        hand_id:
          type: string
    DownloadResponse:
      type: object
      properties:
        clip_request_id:
          type: string
        status:
          type: string
          enum: [queued, processing]
        message:
          type: string
        estimated_completion:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
      description: Next-Auth session cookie (IAP 통합)
security:
  - CookieAuth: []